src/test/java/com/project/crm/controller/UserControllerTest.java
package com.project.crm.controller;

import com.google.gson.Gson;
import com.project.crm.domain.User;
import com.project.crm.domain.UserDto;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import java.util.ArrayList;
import java.util.List;

import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.Matchers.is;
import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(UserController.class)
public class UserControllerTest {
    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserController controller;

    @Test
    public void shouldReturnAllUsers() throws Exception {
        //given
        List<UserDto> users = new ArrayList<>();
        UserDto u1 = new UserDto(1, "maklukas", "fn", "ln", 1);
        users.add(u1);
        String gsonTxt = new Gson().toJson(u1);
        when(controller.getUsers()).thenReturn(users);
        mockMvc.perform(get("/v1/users").contentType(MediaType.APPLICATION_JSON_VALUE))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)))
                .andExpect(jsonPath("$", is(gsonTxt)));
        //when
        //then
    }


}